pyproject.toml
[tool.poetry]
name = "news-terminal"
version = "0.1.0"
description = "Personal news terminal"
authors = ["Thr3ee"]
packages = [{include = "news_terminal", from="src"}]

[tool.poetry.dependencies]
python = "^3.11"
playwright = "^1.30.0"
websockets = "^10.4"
textual = {extras = ["dev"], version = "^0.16.0"}
rich-pixels = "^2.1.1"
unicorn-binance-websocket-api = "^1.42.0"
python-binance = "^1.0.17"
tweepy = {extras = ["async"], version = "^4.13.0"}
python-decouple = "^3.8"
textual-autocomplete = {git = "https://github.com/UmBsublime/textual-autocomplete", rev = "textual_v0.14.0_compat"}


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
